+++ CHANGELOG.rst
@@ -14,6 +14,15 @@ Change Log
 Unreleased
 **********
 
+3.5.0 – 2023-09-04
+**********************************************
+
+* Add edx-drf-extensions lib.
+* Add JwtAuthentication checks before each request.
+* Add SessionAuthentication checks before each request.
+* Add HasStudioWriteAccess permissions checks before each request.
+
+
 3.4.0 – 2023-08-30
 **********************************************